backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cfs-funding-portal"
version = "0.1.0"
description = "CFS Funding Portal + Social Vulnerability Index"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "sqlalchemy[asyncio]>=2.0.35",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "httpx>=0.27.0",
    "msal>=1.31.0",
    "python-jose[cryptography]>=3.3.0",
    "azure-storage-blob>=12.23.0",
    "azure-identity>=1.19.0",
    "structlog>=24.4.0",
    "prometheus-client>=0.21.0",
    "tenacity>=9.0.0",
    "python-multipart>=0.0.12",
    "jinja2>=3.1.4",
    "weasyprint>=62.0",
    "python-docx>=1.1.0",
    "orjson>=3.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "factory-boy>=3.3.0",
    "faker>=30.0.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "A", "SIM", "TCH"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]
